+++ azure/storage/blob/pageblobservice.py
@@ -96,7 +96,7 @@ class PageBlobService(_BaseBlobService):
             timeout, sas_token, connection_string, request_session)
 
     def create_blob(
-        self, container_name, blob_name, content_length, settings=None,
+        self, container_name, blob_name, content_length, content_settings=None,
         sequence_number=None, metadata=None, lease_id=None, if_modified_since=None,
         if_unmodified_since=None, if_match=None, if_none_match=None):
         '''
@@ -114,8 +114,8 @@ class PageBlobService(_BaseBlobService):
             Required. This header specifies the maximum size
             for the page blob, up to 1 TB. The page blob size must be aligned
             to a 512-byte boundary.
-        settings:
-            Settings object used to set properties on the blob.
+        content_settings:
+            ContentSettings object used to set properties on the blob.
         sequence_number:
             The sequence number is a user-controlled value that you can use to
             track requests. The value of the sequence number must be between 0
@@ -152,8 +152,8 @@ class PageBlobService(_BaseBlobService):
             ('If-Match', _str_or_none(if_match)),
             ('If-None-Match', _str_or_none(if_none_match))
         ]
-        if settings is not None:
-            request.headers += settings.to_headers()
+        if content_settings is not None:
+            request.headers += content_settings.to_headers()
 
         request.path, request.query = _update_request_uri_query_local_storage(
             request, self.use_local_storage)
@@ -431,7 +431,7 @@ class PageBlobService(_BaseBlobService):
     #----Convenience APIs-----------------------------------------------------
 
     def create_blob_from_path(
-        self, container_name, blob_name, file_path, settings=None,
+        self, container_name, blob_name, file_path, content_settings=None,
         metadata=None, progress_callback=None, max_connections=1,
         max_retries=5, retry_wait=1.0, lease_id=None, if_modified_since=None,
         if_unmodified_since=None, if_match=None, if_none_match=None):
@@ -445,8 +445,8 @@ class PageBlobService(_BaseBlobService):
             Name of blob to create or update.
         file_path:
             Path of the file to upload as the blob content.
-        settings:
-            Settings object used to set blob properties.
+        content_settings:
+            ContentSettings object used to set blob properties.
         metadata:
             A dict containing name, value for metadata.
         progress_callback:
@@ -485,7 +485,7 @@ class PageBlobService(_BaseBlobService):
                 blob_name=blob_name,
                 stream=stream,
                 count=count,
-                settings=settings,
+                content_settings=content_settings,
                 metadata=metadata,
                 progress_callback=progress_callback,
                 max_connections=max_connections,
@@ -499,7 +499,7 @@ class PageBlobService(_BaseBlobService):
 
 
     def create_blob_from_stream(
-        self, container_name, blob_name, stream, count, settings=None,
+        self, container_name, blob_name, stream, count, content_settings=None,
         metadata=None, progress_callback=None, max_connections=1,
         max_retries=5, retry_wait=1.0, lease_id=None, if_modified_since=None,
         if_unmodified_since=None, if_match=None, if_none_match=None):
@@ -516,8 +516,8 @@ class PageBlobService(_BaseBlobService):
         count:
             Number of bytes to read from the stream. This is required, a page
             blob cannot be created if the count is unknown.
-        settings:
-            Settings object used to set the blob properties.
+        content_settings:
+            ContentSettings object used to set the blob properties.
         metadata:
             A dict containing name, value for metadata.
         progress_callback:
@@ -561,7 +561,7 @@ class PageBlobService(_BaseBlobService):
             container_name=container_name,
             blob_name=blob_name,
             content_length=count,
-            settings=settings,
+            content_settings=content_settings,
             metadata=metadata,
             lease_id=lease_id,
             if_modified_since=if_modified_since,
@@ -587,7 +587,7 @@ class PageBlobService(_BaseBlobService):
 
     def create_blob_from_bytes(
         self, container_name, blob_name, blob, index=0, count=None,
-        settings=None, metadata=None, progress_callback=None,
+        content_settings=None, metadata=None, progress_callback=None,
         max_connections=1, max_retries=5, retry_wait=1.0,
         lease_id=None, if_modified_since=None, if_unmodified_since=None,
         if_match=None, if_none_match=None):
@@ -607,8 +607,8 @@ class PageBlobService(_BaseBlobService):
         count:
             Number of bytes to upload. Set to None or negative value to upload
             all bytes starting from index.
-        settings:
-            Settings object used to set blob properties.
+        content_settings:
+            ContentSettings object used to set blob properties.
         metadata:
             A dict containing name, value for metadata.
         progress_callback:
@@ -655,7 +655,7 @@ class PageBlobService(_BaseBlobService):
             blob_name=blob_name,
             stream=stream,
             count=count,
-            settings=settings,
+            content_settings=content_settings,
             metadata=metadata,
             lease_id=lease_id,
             progress_callback=progress_callback,