+++ azure/storage/file/_chunking.py
@@ -22,7 +22,7 @@ from .._common_serialization import url_quote
 
 class _FileChunkDownloader(object):
     def __init__(self, file_service, share_name, directory_name, file_name, 
-                 file_size, chunk_size, stream, parallel, max_retries, retry_wait,
+                 file_size, chunk_size, stream, max_retries, retry_wait,
                  progress_callback):
         self.file_service = file_service
         self.share_name = share_name
@@ -31,11 +31,11 @@ class _FileChunkDownloader(object):
         self.file_size = file_size
         self.chunk_size = chunk_size
         self.stream = stream
-        self.stream_start = stream.tell() if parallel else None
-        self.stream_lock = threading.Lock() if parallel else None
+        self.stream_start = stream.tell()
+        self.stream_lock = threading.Lock()
         self.progress_callback = progress_callback
         self.progress_total = 0
-        self.progress_lock = threading.Lock() if parallel else None
+        self.progress_lock = threading.Lock()
         self.max_retries = max_retries
         self.retry_wait = retry_wait
 
@@ -54,21 +54,14 @@ class _FileChunkDownloader(object):
 
     def _update_progress(self, length):
         if self.progress_callback is not None:
-            if self.progress_lock is not None:
-                with self.progress_lock:
-                    self.progress_total += length
-                    total = self.progress_total
-            else:
+            with self.progress_lock:
                 self.progress_total += length
                 total = self.progress_total
-            self.progress_callback(total, self.file_size)
+                self.progress_callback(total, self.file_size)
 
     def _write_to_stream(self, chunk_data, chunk_offset):
-        if self.stream_lock is not None:
-            with self.stream_lock:
-                self.stream.seek(self.stream_start + chunk_offset)
-                self.stream.write(chunk_data)
-        else:
+        with self.stream_lock:
+            self.stream.seek(self.stream_start + chunk_offset)
             self.stream.write(chunk_data)
 
     def _download_chunk_with_retries(self, chunk_offset):
@@ -196,6 +189,12 @@ class _FileChunkUploader(object):
 def _download_file_chunks(file_service, share_name, directory_name, file_name,
                           file_size, block_size, stream, max_connections,
                           max_retries, retry_wait, progress_callback):
+    if max_connections <= 1:
+        raise ValueError(
+            'To use file chunk downloader more than 1 thread must be ' +
+            'used since get_blob should be called for single threaded ' +
+            'file downloads.')
+
     downloader = _FileChunkDownloader(
         file_service,
         share_name,
@@ -204,7 +203,6 @@ def _download_file_chunks(file_service, share_name, directory_name, file_name,
         file_size,
         block_size,
         stream,
-        max_connections > 1,
         max_retries,
         retry_wait,
         progress_callback,
@@ -213,14 +211,9 @@ def _download_file_chunks(file_service, share_name, directory_name, file_name,
     if progress_callback is not None:
         progress_callback(0, file_size)
 
-    if max_connections > 1:
-        import concurrent.futures
-        executor = concurrent.futures.ThreadPoolExecutor(max_connections)
-        result = list(executor.map(downloader.process_chunk, downloader.get_chunk_offsets()))
-    else:
-        for range_start in downloader.get_chunk_offsets():
-            downloader.process_chunk(range_start)
-
+    import concurrent.futures
+    executor = concurrent.futures.ThreadPoolExecutor(max_connections)
+    result = list(executor.map(downloader.process_chunk, downloader.get_chunk_offsets()))
 
 def _upload_file_chunks(file_service, share_name, directory_name, file_name,
                         file_size, block_size, stream, max_connections,